+++ js/utils.js
@@ -7,14 +7,10 @@ const getRandomInteger = function (min, max) {
   return Math.floor(Math.random() * (max - min + 1)) + min;
 }
 
-getRandomInteger(0, 10);
-
 //Функция для проверки максимальной длины строки
 const getCommentLength = function (text, maxLength) {
   return text.length <= maxLength;
 };
-
-getCommentLength('Функция для проверки максимальной длины строки', 140)
 
 //функция по поиску случайного элемента в переданном массиве
 const getRandomArrayElement = (elements) => {
